refactor(localize): move `source_file_utils.ts` up
This will allow the utilities in this file to be shared outside
`translate` code.

Some more text to get to the 100 character commit message requirement.
