+++ packages/localize/src/tools/src/source_file_utils.ts
@@ -8,8 +8,7 @@
 import {ɵisMissingTranslationError, ɵmakeTemplateObject, ɵParsedTranslation, ɵtranslate} from '@angular/localize';
 import {NodePath} from '@babel/traverse';
 import * as t from '@babel/types';
-
-import {Diagnostics} from '../../diagnostics';
+import {Diagnostics} from './diagnostics';
 
 /**
  * Is the given `expression` the global `$localize` identifier?
