+++ packages/localize/src/tools/src/translate/main.ts
@@ -10,17 +10,17 @@ import * as glob from 'glob';
 import {resolve} from 'path';
 import * as yargs from 'yargs';
 
+import {Diagnostics} from '../diagnostics';
+import {MissingTranslationStrategy} from '../source_file_utils';
 import {AssetTranslationHandler} from './asset_files/asset_translation_handler';
 import {getOutputPathFn, OutputPathFn} from './output_path';
 import {SourceFileTranslationHandler} from './source_files/source_file_translation_handler';
-import {MissingTranslationStrategy} from './source_files/source_file_utils';
 import {TranslationLoader} from './translation_files/translation_loader';
 import {SimpleJsonTranslationParser} from './translation_files/translation_parsers/simple_json_translation_parser';
 import {Xliff1TranslationParser} from './translation_files/translation_parsers/xliff1_translation_parser';
 import {Xliff2TranslationParser} from './translation_files/translation_parsers/xliff2_translation_parser';
 import {XtbTranslationParser} from './translation_files/translation_parsers/xtb_translation_parser';
 import {Translator} from './translator';
-import {Diagnostics} from '../diagnostics';
 
 if (require.main === module) {
   const args = process.argv.slice(2);
